Refactor version.rs -> standard.rs
