+++ vhdl_lang/src/project.rs
@@ -11,8 +11,8 @@ use crate::completion::{list_completion_options, CompletionItem};
 use crate::config::Config;
 use crate::lint::dead_code::UnusedDeclarationsLinter;
 use crate::named_entity::{AnyEnt, EntRef};
+use crate::standard::VHDLStandard;
 use crate::syntax::VHDLParser;
-use crate::version::VHDLStandard;
 use crate::{data::*, EntHierarchy, EntityId};
 use fnv::{FnvHashMap, FnvHashSet};
 use std::collections::hash_map::Entry;
