+++ vhdl_lang/src/syntax/parser.rs
@@ -8,7 +8,7 @@ use super::design_unit::parse_design_file;
 use super::tokens::{Symbols, TokenStream, Tokenizer};
 use crate::ast::DesignFile;
 use crate::data::*;
-use crate::version::VHDLStandard;
+use crate::standard::VHDLStandard;
 use std::io;
 use std::sync::Arc;
