+++ vhdl_lang/src/syntax/test.rs
@@ -33,12 +33,12 @@ use crate::ast;
 use crate::ast::*;
 use crate::data::Range;
 use crate::data::*;
+use crate::standard::VHDLStandard;
 use crate::syntax::concurrent_statement::parse_map_aspect;
 use crate::syntax::context::{parse_context, DeclarationOrReference};
 use crate::syntax::names::parse_association_element;
 use crate::syntax::subprogram::{parse_optional_subprogram_header, parse_subprogram_instantiation};
 use crate::syntax::{kind_str, TokenAccess, TokenId, TokenSpan};
-use crate::version::VHDLStandard;
 use std::collections::hash_map::DefaultHasher;
 use std::collections::hash_map::Entry;
 use std::collections::HashMap;
